requirements.txt
discord.py
python-dotenv
SQLAlchemy
requests
pydantic-settings